% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra.R
\name{calculate_area}
\alias{calculate_area}
\title{Calculate area under spectral curve}
\usage{
calculate_area(wavelengths, excitations, emissions, lasers, channels,
  threshold = 0.01)
}
\arguments{
\item{wavelengths}{A vector of wavelengths (nm).}

\item{excitations}{A vector of excitation intensities.}

\item{emissions}{A vector of emmision intensities.}

\item{lasers}{Vector of lasers exciting given proteins e.g. c(488, 642).}

\item{channels}{Vector of channels for which spectral area is calculated
e.g. c('530/30', '670LP').}

\item{threshold}{Fraction of total spectral area observed in a given
channel that is considered to be practically 0.}
}
\value{
A data frame with two columns -- channel and area.
}
\description{
Calculates the area under an emission spectra within specified channels,
excited by given laser wavelength. The output is a dataframe to
facilitate use within dplyr's do() function.
}

