% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{calculate_colours}
\alias{calculate_colours}
\title{Generate colour from wavelength}
\usage{
calculate_colours(wavelengths)
}
\arguments{
\item{wavelengths}{A vector of wavelengths (nm) between 300 and 800.}
}
\value{
A vector of colours in hex code notation.
}
\description{
Approximates colour of visual light spectrum.
}

