% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra.R
\name{calculate_max_overlap}
\alias{calculate_max_overlap}
\title{Calculate worst-case overlap between combination of proteins}
\usage{
calculate_max_overlap(wavelengths, excitations, emissions, proteins, n, lasers,
  channels, spectra.threshold = 0.01, intensity.threshold = 0.01)
}
\arguments{
\item{wavelengths}{A vector of wavelengths (nm).}

\item{excitations}{A vector of excitation intensities.}

\item{emissions}{A vector of emmision intensities.}

\item{proteins}{A vector of protein labels that correspond to each
measurement.}

\item{n}{The number of protein combinations to consider.}

\item{lasers}{Vector of lasers exciting given proteins e.g. c(488, 642).}

\item{channels}{Vector of channels for which spectral area is calculated
e.g. c('530/30', '670LP').}

\item{spectra.threshold}{Fraction of total spectral area observed in a
given channel that is considered to be
practically 0.}

\item{channel.threshold}{Fraction of maximum area observed within a given
channel that is considered to be practically 0
(compared to selected protein subset).}

\item{overall.threshold}{Fraction of maximum area observed across all
channels that is considered to be practically 0
(compared to selected protein subset).}
}
\value{
A data frame with two columns -- channel and area.
}
\description{
Calculates the overlaps in area under an emission spectra between all
proteins. Areas are calculated by taking into account the specified
channels, and excitation by given laser wavelength. All combination
of proteins are assessed to determine worst case overlap of one protein
with all others in any given channel. Each protein is assigned one
channel for quantification, with the channels picked in order of least
overlap.
}

