% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra.R
\name{calculate_pairwise_overlap}
\alias{calculate_pairwise_overlap}
\title{Calculate overlap between proteins}
\usage{
calculate_pairwise_overlap(wavelengths, excitations, emissions, proteins,
  lasers, channels, threshold = 0.01)
}
\arguments{
\item{wavelengths}{A vector of wavelengths (nm).}

\item{excitations}{A vector of excitation intensities.}

\item{emissions}{A vector of emmision intensities.}

\item{proteins}{A vector of protein labels that correspond to each
measurement.}

\item{lasers}{Vector of lasers exciting given proteins e.g. c(488, 642).}

\item{channels}{Vector of channels for which spectral area is calculated
e.g. c('530/30', '670LP').}

\item{threshold}{Fraction of total spectral area observed in a given
channel that is considered to be practically 0.}
}
\value{
A data frame with four columns -- protein1, protein2, channel
        and overlap. Overlap represents the fraction of protein1 area
        occupied by protein2 in a given channel.
}
\description{
Calculates the overlaps in area under an emission spectra between all
proteins. Areas are calculated by taking into account the specified
channels, and excitation by given laser wavelength. Overlap is reported
as a fraction of area of specified protein. As an example the overlap
of EGFP by EYFP will likely be different from the overlap of EYFP by EGFP.
An overlap of NA means that the area of specified protein is 0 for the
given channel (and would result in division by 0). An overlap of 0 means
that the relative area of overlapping proteins is practically 0.
}

