% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{calculate_ranges}
\alias{calculate_ranges}
\title{Calculate flow cytometer channel range.}
\usage{
calculate_ranges(channels, min.wavelength, max.wavelength)
}
\arguments{
\item{channels}{Vector of channels e.g. c('530/30', '670LP').}

\item{min.wavelengths}{Minimum wavlength for calculating short pass extent.}

\item{max.wavelengths}{Maximum wavlength for calculating long pass extent.}
}
\value{
A list of c(min, max) ranges.
}
\description{
Converts channel specifications into a list of c(min, max) ranges. Parsing
script inteprets specifications of the form xxxLP for long pass, xxxSP for
short pass, or xxx/yy for bandpass filter.
}

