% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorolive.R
\name{fluorolive}
\alias{fluorolive}
\title{Launch "app" for interactive spectra display.}
\usage{
fluorolive(...)
}
\arguments{
\item{...}{All arguments passed into runApp. See runApp for details.}
}
\description{
A thin wrapper around shiny's runApp() that provides the path to
required server.R and ui.R files.
}

