% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{generate_spectra}
\alias{generate_spectra}
\title{Import fluorescent protein spectra}
\usage{
generate_spectra(filename = NA, destination = NA, basis = seq(300, 800,
  length.out = 200), overwrite = TRUE)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames
containing csv spectra data. If no filename is
provided, the current working directory is assumed
to be the root directory of the package source and
all csv files are matched in ./inst/protein_spectra.}

\item{destination}{Filename to save the generated Rdata file. If no
destination is provided, the current working
directory is assumed to be the root directory of
the package source and a protein_spectra.Rdata 
file is saved to ./data.}

\item{basis}{A sequence of wavelength values (with the same units
as the data in the csv files) to evaluate all spectra
on.}

\item{overwrite}{TRUE to overwirte existing Rdata file.}
}
\description{
Imports fluorescent protein data, interpolates the available points,
and generates new values using a common basis.
}
\examples{
# To do
}

