% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_cary_data}
\alias{import_cary_data}
\title{Import numeric data from csv output from Cary Eclipse Spectrophotometer}
\usage{
import_cary_data(filename)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}
}
\value{
A data.frame with the following columns: filename, 
        sample, cell, scan, label, excitation, emission, intensity
}
\description{
Imports numeric csv output from a Cary Eclipse Spectrophotometer and 
converts it into long format.
}
\examples{
# To do
}

