% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_cary_meta}
\alias{import_cary_meta}
\title{Import sample description metadata from csv output from 
Cary Eclipse Spectrophotometer}
\usage{
import_cary_meta(filename, aggregate = FALSE)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}

\item{aggregate}{TRUE if metadata should be reported per sample
(aggregating across the excitation sweep) or per single
scan.}
}
\value{
A data.frame with columns of scan parameters.
}
\description{
Imports sample description metadata from csv output from a Cary Eclipse 
Spectrophotometer and converts it into long format.
}
\examples{
# To do
}

