% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_data}
\alias{import_data}
\title{Import numeric data from spectrophotometer output.}
\usage{
import_data(filename, instrument = "cary", ...)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}

\item{...}{Arguments passed to specific instrument import functions.}

\item{instrumet}{String specifying instrument. Currently limited to
'Cary Eclipse' or 'cary', 'BioTek Synergy 4' or 
'synergy4'.}
}
\value{
A data.frame with columns of scan parameters.
}
\description{
<<<<<<< HEAD
Imports numeric data from exported output and converts it 
into long format. Import procedure depends on specified instrument. 
Calls import_cary_data() or import_synergy4_data().
=======
Imports numeric csv output from a Cary Eclipse Spectrophotometer and 
converts it into long format.
>>>>>>> shiny
}
\examples{
# To do
}

