% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_meta}
\alias{import_meta}
\title{Import sample description metadata from spectrophotometer output.}
\usage{
import_meta(filename, instrument = "cary", ...)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}

\item{...}{Arguments passed to specific instrument import functions.}

\item{instrumet}{String specifying instrument. Currently limited to
'Cary Eclipse' or 'cary'.}
}
\value{
A data.frame with columns of scan parameters.
}
\description{
Imports sample description metadata from exported output and converts it 
into long format. Import procedure depends on specified instrument. 
[This function is a placeholder as only Cary Eclipse import is currently
supported]. Calls import_cary_meta().
}
\examples{
# To do
}

