% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{import_metadata}
\alias{import_metadata}
\title{Import sample description metadata from csv output from
Cary Eclipse Spectrophotometer}
\usage{
import_metadata(filename, aggregate = FALSE, cache = FALSE)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}

\item{aggregate}{TRUE if metadata should be reported per sample
(aggregating across the excitation sweep) or per single
scan.}

\item{cache}{TRUE to cache processed files. Processed files are
recorded by a hash of their contents, ignoring file names
and ensuring a new import if contents change. [NOT IMPLEMENTED]}
}
\value{
A data.frame with columns of scan parameters.
}
\description{
Imports sample description metadata from csv output from a Cary Eclipse
Spectrophotometer and converts it into long format.
}
\examples{
# To do
}

