% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_synergy4_data}
\alias{import_synergy4_data}
\title{Import numeric data from table output of BioTek Synergy 4}
\usage{
import_synergy4_data(filename)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}
}
\value{
A data.frame with the following columns: filename, read, 
        well, row, column, excitation, emission, intensity
}
\description{
Imports numeric table output from a BioTek Synergy 4 Spectrophotometer and 
converts it into long format.
}
\examples{
# To do
}

