% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{matrix_to_spectra}
\alias{matrix_to_spectra}
\title{Convert excitation/emission matrix to excitation/emission curves}
\usage{
matrix_to_spectra(excitation, emission, intensity, wavelengths = NA,
  basis = seq(300, 800, length.out = 200))
}
\arguments{
\item{excitation}{A vector of excitations.}

\item{emission}{A vector of emissions.}

\item{intensity}{A vector of intensities corresponding to each
excitation/emission pair.}

\item{basis}{A sequence of wavelength values (in nm) at which
the excitation and emission curves will be evaluated.}
}
\value{
A data.frame with three columns --  wavelength, excitation, 
        and emission (corresponding to the excitation and emission
        curves).
}
\description{
Converts intesity data from an excitation/emission profile. First,
the wavelength of maximum peak intensity is identified at each 
excitation. The relative intensity at maximum emission forms the
excitations curve, while the relative intensity at peak excitation
forms the emission curve.
}

