% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_emissions}
\alias{plot_emissions}
\title{Plot emission spectra as excited by specified laser.}
\usage{
plot_emissions(wavelengths, excitations, emissions, proteins, lasers,
  channels = NA, palette = NA, background.alpha = 0.1)
}
\arguments{
\item{wavelengths}{A vector of wavelengths (nm).}

\item{excitations}{A vector of excitation intensities.}

\item{emissions}{A vector of emmision intensities.}

\item{lasers}{Vector of lasers exciting given proteins e.g. c(488, 642)}

\item{channels}{Optional vector of channels to highlight
e.g. c('530/30', '670LP').}

\item{palette}{Name of colorbrewer palette for discrete scale (if NA,
spectra are coloured by emission wavelength.}

\item{background.alpha}{Transparency value for channel highlight.}

\item{protein}{A vector of protein labels that correspond to each
measurement.}
}
\value{
ggplot2 plot object.
}
\description{
Generates a ggplot2 with emission spectra generated by specified laser.
}

