% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Plot emission and excitation spectra.}
\usage{
plot_spectra(wavelengths, excitations, emissions, proteins, lasers = NA,
  channels = NA, palette = NA, background.alpha = 0.1)
}
\arguments{
\item{wavelengths}{A vector of wavelengths (nm).}

\item{excitations}{A vector of excitation intensities.}

\item{emissions}{A vector of emmision intensities.}

\item{proteins}{A vector of protein labels that correspond to each
measurement.}

\item{lasers}{Optional vector of lasers to highlight e.g. c(488, 642)}

\item{channels}{Optional vector of channels to highlight 
e.g. c('530/30', '670LP').}

\item{palette}{Name of colorbrewer palette for discrete scale (if NA,
spectra are coloured by emission wavelength.}

\item{background.alpha}{Transparency value for channel highlight.}
}
\value{
ggplot2 plot object.
}
\description{
Generates a ggplot2 plot with an overview of excitation and 
emission spectra.
}

