% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_data}
\alias{import_data}
\title{Import numeric data from csv output from Cary Eclipse Spectrophotometer}
\usage{
import_data(filename, cache = FALSE)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}

\item{cache}{TRUE to cache processed files. Processed files are
recorded by a hash of their contents, ignoring file names
and ensuring a new import if contents change. [NOT IMPLEMENTED]}
}
\value{
A data.frame with the following columns: sample, cell, scan,
        label, excitation, emission, intensity
}
\description{
Imports numeric csv output from a Cary Eclipse Spectrophotometer and 
converts it into long format.
}
\examples{
# To do
}

