% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fluoroscripts.R
\docType{data}
\name{protein.spectra}
\alias{protein.spectra}
\title{Spectral information of selected fluorescent proteins.}
\format{A data frame with 4800 rows and 6 variables.
\describe{
   \item{protein}{Name of protein}
   \item{wavelength}{Wavelength in nm}
   \item{r.excitation}{Relative excitation curve scaled to range of 0-1}
   \item{r.emission}{Relative emission curve scaled to range of 0-1}
   \item{excitation}{Relative excitation curve multiplied by brightness}
   \item{emission}{Relative emission curve multiplied by brightness}
}}
\usage{
protein.spectra
}
\description{
Information combines spectra available through
http://www.spectra.arizona.edu/ with manual spectra plot digitation from
original publications.
}
\keyword{datasets}

